+++ zcash_client_sqlite/src/testing/db.rs
@@ -38,6 +38,7 @@ use zcash_primitives::{
 };
 use zcash_protocol::{
     ShieldedPool, consensus, consensus::BlockHeight, local_consensus::LocalNetwork, memo::Memo,
+    value::Zatoshis,
 };
 use zip32::DiversifierIndex;